Expand items

+++ include/player_pc.h
@@ -8,7 +8,7 @@ struct PlayerPCItemPageStruct
     u16 cursorPos;
     u16 itemsAbove;
     u8 pageItems;
-    u8 count;
+    u16 count;
     u8 filler[3];
     u8 scrollIndicatorTaskId;
 };
